adversarial_playground/src/data_resnet/requirements.txt
google-cloud-storage
kagglehub